auth/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.7.2
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
black==24.2.0
    # via -r requirements.in
build==1.0.3
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2024.2.2
    # via
    #   -r requirements.in
    #   requests
cffi==1.16.0
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
cryptography==42.0.5
    # via
    #   -r requirements.in
    #   jwcrypto
    #   pyjwt
defusedxml==0.7.1
    # via
    #   -r requirements.in
    #   python3-openid
django==5.0.2
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-cors-headers
    #   django-oauth-toolkit
    #   djangorestframework
django-allauth==0.61.1
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-oauth-toolkit==2.3.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
djangorestframework-oauth==1.1.0
    # via -r requirements.in
idna==3.6
    # via
    #   -r requirements.in
    #   requests
isort==5.13.2
    # via -r requirements.in
jwcrypto==1.5.4
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
kafka-python==2.0.2
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   django-oauth-toolkit
    #   requests-oauthlib
packaging==23.2
    # via
    #   -r requirements.in
    #   black
    #   build
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   black
pip-tools==7.4.0
    # via -r requirements.in
platformdirs==4.2.0
    # via
    #   -r requirements.in
    #   black
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pyjwt[crypto]==2.8.0
    # via
    #   -r requirements.in
    #   django-allauth
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python3-openid==3.2.0
    # via
    #   -r requirements.in
    #   django-allauth
pytz==2024.1
    # via
    #   -r requirements.in
    #   djangorestframework
requests==2.31.0
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-oauth-toolkit
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   -r requirements.in
    #   django-allauth
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
typing-extensions==4.10.0
    # via
    #   -r requirements.in
    #   jwcrypto
ujson==5.9.0
    # via -r requirements.in
urllib3==2.2.1
    # via
    #   -r requirements.in
    #   requests
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
